[Interfaces] Moved AudioGenerator and EmbeddingsGenerator to max.interfaces
MODULAR_ORIG_COMMIT_REV_ID: 96a480cf884d979039536a0a64f3edc20a985a47
