--- max/pipelines/core/interfaces/__init__.py
@@ -1,24 +1,0 @@
-# ===----------------------------------------------------------------------=== #
-# Copyright (c) 2025, Modular Inc. All rights reserved.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions:
-# https://llvm.org/LICENSE.txt
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ===----------------------------------------------------------------------=== #
-
-from .audio_generation import (
-    AudioGenerator,
-    AudioGeneratorContext,
-)
-from .embeddings_generation import EmbeddingsGenerator
-
-__all__ = [
-    "AudioGenerator",
-    "AudioGeneratorContext",
-    "EmbeddingsGenerator",
-]
