+++ max/pipelines/lib/embeddings_pipeline.py
@@ -21,11 +21,8 @@ from typing import TYPE_CHECKING, Any, TypeVar
 from max.driver import load_devices
 from max.engine import InferenceSession
 from max.graph.weights import load_weights
-from max.interfaces import EmbeddingsOutput, InputContext
+from max.interfaces import EmbeddingsGenerator, EmbeddingsOutput, InputContext
 from max.nn import ReturnLogits
-from max.pipelines.core import (
-    EmbeddingsGenerator,
-)
 from max.profiler import Tracer, traced
 
 if TYPE_CHECKING:
