+++ max/serve/api_server.py
@@ -23,9 +23,8 @@ from typing import Union
 
 from fastapi import FastAPI, Response
 from fastapi.responses import JSONResponse
-from max.interfaces import PipelineTask, PipelineTokenizer
+from max.interfaces import PipelinesFactory, PipelineTask, PipelineTokenizer
 from max.nn.kv_cache import KVTransferEngineMetadata
-from max.pipelines.core import PipelinesFactory
 from max.pipelines.lib import PipelineConfig
 from max.serve.config import APIType, MetricRecordingMethod, Settings
 from max.serve.kvcache_agent.dispatcher_factory import DispatcherFactory
